Add snmalloc as a sanitizer option
This commit adds snmalloc as an option for the sanitizer.
It requires the snmalloc code to be checked out and built.
The following sequence of commands starting in the root of the checkout
should work.
```
cd sanitizers
git clone https://github.com/microsoft/snmalloc
mkdir snmalloc/build
cd snmalloc/build
cmake ..
make snmallocshim-checks
```

Some of the checks in snmalloc are probablistic, so multiple runs should
be performed to get an accurate representation of its detection rate.